heka/frontend/requirements.txt
geopandas==0.7.0
numpy==1.18.1
plotly==4.5.4
dash==1.9.1
dash_bootstrap_components==0.7.1
dash_html_components==1.0.1
dash_core_components==1.8.1
geopy==1.22.0
pandas==1.0.3
PyYAML==5.3.1
Flask-CSV==1.2.0
Flask==1.1.1
Flask-Caching==1.8.0
requests==2.22.0
pathlib2==2.3.4
multiprocess==0.70.9
gunicorn==20.0.4
brain-isotopic-distribution==1.5.4
pymzml==2.4.7
google-api-python-client==1.10.0
google-cloud==0.34.0
google-cloud-vision==1.0.0
google-cloud-storage==1.30.0
psycopg2==2.8.5
scipy==1.5.2
dash_table==4.4.0
dash-table-experiments==0.6.0